Organize User module files
* Make folder for `Executors` and move executor files into this folder
* Put `OSUserExecutor` class into its own file, removed from the `OSUserRequests` mega-file
* Make folder for `Requests` and move request-related files to this folder
* Split up each request from `OSUserRequests` mega-file into its own file
* Now `OSUserRequest.swift` will contain just the protocol and an extension of OneSignalRequest
